Partially merged brAthena/FluxCP@cb3a0fa
Prevents a possible XSS exploit and revalidates email and username on change/reset.

I did not merge the email regex, because as @MishimaHaruna correctly stated on the hercules merge, this email validation is far to specific and incorrect.

Additionally I also did not merge the other username regex changes, since they would now only partially check the string and not the whole string anymore.

If someone likes he or she is free to merge the maximum checks for password's character amount.

Thanks to @TidusBR for the original commit and to @Mysteries for raising my attention to it by merging it to hercules' flux.