Fixing xss vulnerability on servicedesk tickets
This pull request fixes an XSS vulnerability in the project's code. The vulnerability was allowing malicious code to be injected into a page and executed in the user's browser, potentially stealing their cookies and accessing administrator accounts.

To fix this vulnerability, it was identified that the addslashes() method in PHP was not working correctly. It was changed to the htmlentities() method to ensure that the code is protected from XSS attacks.