Added cookie restrictions (#408)
Bumped PHP minimum version to 7.3.0
This was required to have the new function signature available, which allows setting the samesite attribute natively
PHP 7.3.0 was introduced in 2018 - 7 years ago - it is already out of support since 2021 - 4 years ago.
Updating to a more recent PHP version is recommended.

Bumped recommended PHP and MySQL version to latest available version

This update is aimed to resolve a security vulnerability reported by @marksocrates1111